app/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
mysql-connector-python==8.0.29
